Introduce new OutputCImg template class
This class encapsulates the creation of an Rcpp vector, and a
corresponding CImg instance sharing the vector's buffer, which allows us
to perform operations on the CImg instance and then return the vector to
R without the need for a further memory allocation + copy.

While in principle we can make this class more integrated with Rcpp via
as/wraps overloading, this adds a bit of complexity that isn't needed at
the moment, when we're still experimenting with this.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>