Avoid temporary allocations during colour conversions
The different arithmetic operators defined in CImg work as usual: they
have an in-place version returning a reference to the same object, and a
copy version returning a copy of the modified object. The latter was
being used throughout the colourspace functions when
multiplying/dividing by 255, leading to an extra memory allocation in
each case that wasn't necessary.

This commit changes how these multiplications/divisions are performed so
they happen in place on the output CImg object instead of producing a
new copy.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>