Port initial two routines to new allocation-saving infrastructure
The two first candidates for testing the new infrastructure that saves
us a memory copy are RGBtoHSV, because it's the one that triggered the
observation about the unnecessary allocations, and warp, because that's
one of our own main work horses, and therefore it's in our benefit to
ensure it works efficiently.

Note that also on top of saving a memory allocation for the output
image, warp also doesn't allocate a CImg for the input warpfield either,
which is another nice big.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>