CImg: Avoid potential null dereferences
These were diagnosed as potential issues by gcc 13 and 14. Note however
that the one around line 58255 seems to have been a false positive by
gcc 13, since it's not noted by gcc 14, and the same call in other
places of the function (with the argument not actually changing in
between) were not diagnosed by gcc 13.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>